requirements.txt
flask
flask-cors
flask-restful
python-dotenv
psycopg2-binary
flask-sqlalchemy
Flask-Migrate